+++ utils/join/benefitsData.js
@@ -1,0 +1,52 @@
+const benefitsData = [
+  {
+    title: "Medical, dental and vision",
+    description:
+      "We offer competitive medical, dental, vision insurance for employees and dependents. This includes medical, dental, and vision premiums.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_B_icon_benefits_03_health.svg?a974a5f44",
+    imageAlt: "Icon representing health benefits",
+  },
+  {
+    title: "Diversity, Equity, and Inclusion",
+    description:
+      "We follow DEI policy that promotes fair representation, equal opportunities, and respectful treatment of individuals from different backgrounds, ensuring a diverse workforce or community, fostering an inclusive environment, and striving for equitable outcomes for all.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_B_icon_benefits_01_DEI.svg?510379e31",
+    imageAlt: "Benefits icon showcasing diversity, equity, and inclusion",
+  },
+  {
+    title: "Wellbeing",
+    description:
+      "Resources, workshops & events to support your wellbeing journey: Mind, Body & Soul.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_B_icon_benefits_07_wellbeing.svg?a974a5f44",
+    imageAlt: "Illustration of a flower representing wellbeing",
+  },
+  {
+    title: "Monthly stipend",
+    description:
+      "We offer everyone a monthly stipend to use toward services that fit their personal needs. This gives you the opportunity to choose your own adventure and use the funds for whatever matters most to you.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_B_icon_benefits_01_DEI.svg?510379e31",
+    imageAlt: "Benefits icon showcasing diversity, equity, and inclusion",
+  },
+  {
+    title: "Parental leave",
+    description:
+      "We offer biological, adoptive, and foster parents paid time off to spend quality time with family.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_B_icon_benefits_06_parental.svg?a974a5f44",
+    imageAlt: "Parental leave icon",
+  },
+  {
+    title: "Time off",
+    description:
+      "We want you to take time off to rest and rejuvenate. Notion offers flexible paid vacation as well as 10+ observed holidays by country.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_B_icon_benefits_05_pto.svg?a974a5f44",
+    imageAlt: "Benefits icon depicting paid time off",
+  },
+];
+
+export default benefitsData;
