+++ utils/join/communityData.js
@@ -1,0 +1,28 @@
+const communityData = [
+  {
+    title: "Community",
+    description:
+      "Community and culture are paramount at Informatician. We curate meaningful moments to bring employees together for connection, collaboration, and celebration. Whether it's spending the day giving back philanthropically, channeling your inner entrepreneur at our annual Hackweek, showing off your hidden skills at the Informatician talent show, or joining one of Informatician's Employee Resource Groups, there is something for everyone.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_E_img_01_@2x.jpg?a974a5f44",
+    imageAlt: "A group of people excited",
+  },
+  {
+    title: "Characters with character",
+    description:
+      "We believe diverse stories and backgrounds are what make Informatician great. From candidates all the way to extending offers, our hiring practices prioritize diversity, equity, and inclusion. We have company-wide goals to hire more folks from underrepresented groups or that have unique experiences!",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_E_img_02_@2x.jpg?a974a5f44",
+    imageAlt: "Three people staring at you",
+  },
+  {
+    title: "Growth Mentality",
+    description:
+      "We believe that ongoing training and development bring out the best in our employees and our company as a whole. We're constantly researching and implementing opportunities to help our people grow and thrive along their career journey.",
+    imageSrc:
+      "https://s-f.scribdassets.com/images/about/03_careers_E_img_03_@2x.jpg?a974a5f44",
+    imageAlt: "A girl sitting and reading a book",
+  },
+];
+
+export default communityData;
