+++ utils/join/employeeData.js
@@ -1,0 +1,17 @@
+const employeesData = [
+  {
+    name: "Rohan Sharma",
+    imageSrc: "https://avatars.githubusercontent.com/u/33249782?v=4",
+    testimonial:
+      "The vibrant and dynamic environment fosters creativity and innovation. Being part of a talented team that is passionate about improving access to knowledge and reading has been incredibly fulfilling.",
+  },
+  {
+    name: "Deepak",
+    imageSrc:
+      "https://images.weserv.nl/?output=webp&url=https://avatars.githubusercontent.com/u/96341431?v=4",
+    testimonial:
+      "If you're motivated and persistent, your impact will definitely be felt by the company. There is nothing more rewarding than seeing your hard work turned into something fruitful that directly helps the company.",
+  },
+];
+
+export default employeesData;
