+++ utils/join/positionsData.js
@@ -1,0 +1,32 @@
+const positionsData = [
+  {
+    title: "Data Science & Machine Learning",
+    count: 3,
+  },
+  {
+    title: "Product, Designs, & Analytics",
+    count: 2,
+  },
+  {
+    title: "People",
+    count: 0,
+  },
+  {
+    title: "Engineering",
+    count: 2,
+  },
+  {
+    title: "Marketing",
+    count: 3,
+  },
+  {
+    title: "Finance and Operations",
+    count: 5,
+  },
+  {
+    title: "Content Acquisition",
+    count: 1,
+  },
+];
+
+export default positionsData
